# Faster strm exports