Google Drive Test Directory:
```
Export Directory:
    /Movies/

Contents:
    /Movies/Movie1/Movie1.mp4
    /Movies/Movie2/Movie2.mp4
    /Movies/Movie3/Movie3.mp4
    ... 2,292 total movies
```

**Before Changes**
Google Drive API Requests: 2317
Export Duration: 15 minutes and 40 seconds (940 seconds)

**After Changes**
Google Drive API Requests: 28
Export Duration: 4 minutes and 26 seconds (266 seconds)