@0o120 2023-02-02
Was able to get that same 266 second export down to 47 seconds by disabling the export_manager database access inside the `while` loop located inside `clouddrive.common.service.export.ExportService.process_pending_changes` and saving the items after the loop.

You can see in these changes:
https://github.com/0o120/script.module.clouddrive.common-1/commit/c5a5b7b7c10dff96992cc559fff06718ac62d8c2

I'm assuming export_manager.save_pending_changes() is used to save progress in the event of an interruption and not exactly necessary? Maybe adding a config option to disable save_pending_changes inside that loop in exchange for performance gain?

I'm up for adding a config option. Some thoughts on this would be great :)