@cguZZman 2024-04-21
Thanks for the improvement. It's been a long time and this project is been in KTLO mode for me. But I will merge your change if you confirm it's been working fine for you. 
Also, saving the progress is necessary for large catalogs. If you submit your other change with a config option, I could merge it too.